Move withSearchIndex into it's own file
